requirements.txt
absl-py==0.7.1
appdirs==1.4.3
argh==0.26.2
astor==0.7.1
astunparse==1.6.3
atomicwrites==1.2.1
attrs==18.1.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
cachetools==4.0.0
certifi==2017.7.27.1
chardet==3.0.4
cleverhans==2.1.0
Click==7.0
cloudpickle==1.3.0
configparser==4.0.2
contextlib2==0.5.5
coverage==5.1
cplex==12.8.0.0
cycler==0.10.0
dask==1.1.5
decorator==4.3.0
deepdish==0.3.6
defusedxml==0.5.0
dill==0.3.1.1
docker-pycreds==0.4.0
docloud==1.0.257
docplex==2.4.61
edward==1.3.5
efficientnet==1.1.0
emnist==0.0
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
gast==0.3.3
gitdb2==2.0.6
GitPython==3.0.5
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
gpytorch==0.1.0rc4
gql==0.2.0
graphql-core==1.1
graphviz==0.10.1
grpcio==1.27.2
h5py==2.10.0
html5lib==0.9999999
idna==2.6
imageio==2.5.0
imgaug==0.2.8
infinity==1.4
interval==1.0.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
joblib==0.12.5
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
llvmlite==0.25.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.4
mnist==0.2.2
mock==2.0.0
more-itertools==4.3.0
mss==3.3.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.5
nose==1.3.7
notebook==5.7.6
numba==0.40.0
numexpr==2.6.6
numpy==1.16.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.0.0.21
opt-einsum==2.3.2
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.4
pathlib2==2.3.2
pathtools==0.1.2
pbr==4.1.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
plotly==4.5.3
pluggy==0.7.1
prettyprint==0.1.5
prometheus-client==0.6.0
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.11.3
psutil==5.4.6
ptyprocess==0.6.0
py==1.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclustering==0.9.3.1
pycodestyle==2.4.0
pydot==1.2.4
Pygments==2.2.0
pymap3d==1.6.3
pyparsing==2.3.1
pyro-ppl==0.3.0
pyrsistent==0.14.11
pyswarm==0.6
pyswarms==1.0.2
pytest==3.8.2
python-dateutil==2.7.5
python-intervals==1.8.0
pytictoc==1.5.0
pytools==2018.5.2
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.12
pyzmq==18.0.1
qtconsole==4.4.3
regex==2020.5.14
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scandir==1.7
scikit-image==0.14.2
scikit-learn==0.19.2
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
sentry-sdk==0.14.1
Shapely==1.6.4.post2
shortuuid==0.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smmap2==2.0.5
subprocess32==3.5.4
swifter==0.153
tables==3.4.4
tb-nightly==1.14.0a20190301
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-datasets==3.1.0
tensorflow-estimator==2.2.0
tensorflow-gan==2.0.0
tensorflow-gpu==2.2.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.22.0
tensorflow-probability==0.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
Theano==1.0.2
tinydb==3.11.1
toolz==0.9.0
torch==1.0.0
tornado==6.0.2
tqdm==4.46.0
traitlets==4.3.2
urllib3==1.22
wandb==0.8.21
watchdog==0.9.0
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.2
